Rename calico typhe metrics patch file
