+++ kubemarine/patches/__init__.py
@@ -22,7 +22,7 @@ The whole directory is automatically cleared and reset after new version of Kube
 from typing import List
 
 from kubemarine.core.patch import Patch
-from kubemarine.patches.p2_calico_typha_metrics import EnableCalicoTyphaMetrics
+from kubemarine.patches.p3_calico_typha_metrics import EnableCalicoTyphaMetrics
 from kubemarine.patches.p2_enable_calico_audit import EnableCalicoAudit
 from kubemarine.patches.p1_api_version import InsertAPIVersion
